+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added list-view link to user in BillingAgreement admin
+- Search more fields in BillingAgreement admin (user id, username, email)
 
 ## [0.15.3] - 2020-03-25
 ### Fixed